refactor lockfile drift seams
